add StandardWellEval
