+++ opm/simulators/wells/StandardWellGeneric.cpp
@@ -48,12 +48,11 @@ namespace Opm
 template<class Scalar>
 StandardWellGeneric<Scalar>::
 StandardWellGeneric(int Bhp,
-                    const ParallelWellInfo& pw_info,
                     const WellInterfaceGeneric& baseif)
     : baseif_(baseif)
     , perf_densities_(baseif_.numPerfs())
     , perf_pressure_diffs_(baseif_.numPerfs())
-    , parallelB_(duneB_, pw_info)
+    , parallelB_(duneB_, baseif_.parallelWellInfo())
     , Bhp_(Bhp)
 {
     duneB_.setBuildMode(OffDiagMatWell::row_wise);
